fix up types for unchained in `procRes` (single arg overload)
These are also required to not have to treat the derivatives of being
of types `T²`, i.e. a square of elements of type `T`. Technically that
is the case, but it's not needed. If one could implement that
properly, it'd be nice, but I don't have the patience right now to
handle that.