Added tests for power and log + unchained suite
as I failed generate meaningful fixes today, I've generated a longer test suite, that includes "pure" tests for power and log, and "unchained" version for all "pure" tests.
I made this to get a better picture of what is currently working ad what is not at the moment. I hope this helps.